Continued implementation of Matrix class.
Former-commit-id: 1251cc3f434124b158a13bd6dc74a7e054a20c9c

+++ raytracerLib/Matrix.h
@@ -43,10 +43,10 @@ public:
 
 
 	/**
-	 * Returns the first nonzero value in the row.
-	 * Returns 0.0 if all of the entries in the row are zero.
+	 * Returns the column of the first nonzero value in the row.
+	 * Returns -1 if all of the entries in the row are zero.
 	 */
-	double FindFirstNonzeroValue() const;
+	int FindFirstNonzeroValue() const;
 
 
 	/**
